tail.c: remove signed/unsigned comparisons
ext2_off_t is defined as a (usually signed) "int", while
struct ext2_inode member i_size is defined as "__u32".

We make sure the comparisons between the two types are
clean in the sense that we make sure i_size is in the
range shared between the signed and unsigned 32bit integer
type.

These checks are only guaranteed to work if the sizes of
the types match, so we add some macros to define compile
time assertions which we can compare the sizes with.